+++ service/controller/resource/monitoring/prometheus/resource.go
@@ -182,18 +182,6 @@ func toPrometheus(ctx context.Context, v interface{}, config Config) (metav1.Obj
 						},
 					},
 				},
-				// Add 5mn initial delay to the readinessProbe to give Prometheus more
-				// time to load data and start. This fixes the current 10mn delay set
-				// by default in prometheus-operator which results in crash looping Prometheus
-				// when there are too much data to load.
-				Containers: []corev1.Container{
-					{
-						Name: "prometheus",
-						ReadinessProbe: &corev1.Probe{
-							InitialDelaySeconds: 300,
-						},
-					},
-				},
 				EnableFeatures: []string{"remote-write-receiver"},
 				ExternalLabels: map[string]string{
 					key.ClusterIDKey:    key.ClusterID(cluster),